chore: club intro 컴포넌트로 분리
